public/api/pronostico/python/Constanza_v15/requirements.txt
accelerate==0.23.0
anyio==4.0.0
appdirs==1.4.4
attrs==23.1.0
backoff==2.2.1
boilerpy3==1.0.6
cattrs==23.1.2
certifi==2023.7.22
charset-normalizer==3.3.1
click==8.1.7
construct==2.5.3
dnspython==2.4.2
docopt==0.6.2
Events==0.5
exceptiongroup==1.1.3
farm-haystack==1.21.2
filelock==3.12.4
fsspec==2023.10.0
gTTS==2.4.0
h11==0.14.0
haystack==0.42
httpcore==0.18.0
httpx==0.25.0
huggingface-hub==0.18.0
idna==3.4
inflect==7.0.0
Jinja2==3.1.2
joblib==1.3.2
jsonschema==4.19.1
jsonschema-specifications==2023.7.1
lazy-imports==0.3.1
MarkupSafe==2.1.3
monotonic==1.6
more-itertools==10.1.0
mpmath==1.3.0
networkx==3.2
nltk==3.8.1
num2words==0.5.13
numpy==1.26.1
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.18.1
nvidia-nvjitlink-cu12==12.3.52
nvidia-nvtx-cu12==12.1.105
packaging==23.2
pandas==2.1.1
pefile==2023.2.7
Pillow==10.1.0
platformdirs==3.11.0
posthog==3.0.2
prompthub-py==4.0.0
protobuf==4.24.4
psutil==5.9.6
pydantic==1.10.13
pymongo==4.5.0
python-dateutil==2.8.2
python-dotenv==1.0.0
python-ptrace==0.9.8
pytz==2023.3.post1
PyYAML==6.0.1
quantulum3==0.9.0
rank-bm25==0.2.2
referencing==0.30.2
regex==2023.10.3
requests==2.31.0
requests-cache==0.9.8
rpds-py==0.10.6
safetensors==0.4.0
scikit-learn==1.3.2
scipy==1.11.3
sentence-transformers==2.2.2
sentencepiece==0.1.99
six==1.16.0
sniffio==1.3.0
sseclient-py==1.8.0
sympy==1.12
tenacity==8.2.3
threadpoolctl==3.2.0
tiktoken==0.5.1
tokenizers==0.13.3
torch==2.1.0
torchvision==0.16.0
tqdm==4.66.1
transformers==4.32.1
triton==2.1.0
typing_extensions==4.8.0
tzdata==2023.3
url-normalize==1.4.3
urllib3==2.0.7
